feat : Car 객체 이름 원시값 포장
Name 포장 객체에서 이름 규칙에 맞는 검증을 해준다.
car.getName.getName() 처럼 무분별한 getter 사용을 하지 않기 위해, Car의 Name 객체를 활용해 메서드를 만들어준다
