+++ src/main/java/racingcar/domain/car/Car.java
@@ -5,18 +5,13 @@ import racingcar.domain.numbergenerator.NumberGenerator;
 import java.util.Objects;
 
 public class Car {
-    private static final int MAX_NAME_LENGTH = 5;
-    private static final String NAME_LENGTH_ERROR = "[ERROR] 이름은 5글자를 초과할 수 없습니다.";
-    private static final String EMPTY_NAME_ERROR = "[ERROR] 이름은 공백일 수 없습니다.";
     private static final int MOVE_CRITERIA = 4;
 
-    private final String name;
+    private final Name name;
     private int position;
 
     public Car(String name) {
-        validateEmptyName(name);
-        validateName(name);
-        this.name = name;
+        this.name = new Name(name);
         position = 0;
     }
 
@@ -32,20 +27,9 @@ public class Car {
         }
     }
 
-    private void validateEmptyName(String name) {
-        if (name.isEmpty()) {
-            throw new IllegalArgumentException(EMPTY_NAME_ERROR);
-        }
-    }
-
-    private void validateName(String name) {
-        if (name.length() > MAX_NAME_LENGTH) {
-            throw new IllegalArgumentException(NAME_LENGTH_ERROR);
-        }
-    }
 
     public String getName() {
-        return this.name;
+        return name.getName();
     }
 
     public int getPosition() {
